Added 'statbuf' call to the FS interface. Added gc-avoiding refcounts
Added a first stab at ensuring the OCaml functions that get passed to
libuv don't get gc'd. This has no tests yet and is just to get something
up there to talk about.

Not sure we'll keep the statbuf interface this way.
We might end up just passing the stat structure to the FS calls that
actually use it, as opposed to allowing every FS type to be queriable
but return a useless object.

Cleaned up things a bunch more (more struct names have underscores, and
deleted a bunch of not useful comments). Things are still a mess though.

Loop and callback arguments are now both optional. This makes it nicer to
program with, and exposes the sync functionality of libuv to the programer:
if one passes None to the callback the libuv function will be synchronous.

Examples and tests updated.