Commit of a working tcp_echo_server
This is an example based upon the tcp-echo-server in the
libuv book.

Lots of dirty code. Initial stab we'll clean up soon.

This demonstrates a Handle, the subtyping relationship
between Handle -> Stream -> TCP, and a write request.

The callback-gc-avoidance mechanism needs to be abstracted and tested.
Right now there's a ton of repeat code in there that may or may not work.

Also: standardized Loop's run method to take optional loop param