WIP Add example accessor function
We're going to change how we interact with the libuv structs:
instead of specifying the struct fields to determine memory layout
we will determine the size of the struct and then allocate that many
bytes. This will allow us to make less changes as libuv changes.
Also it is helpful because it is cross platform.

However, if we do this, we need a way to access the public struct
fields. This commit shows the way we'll be doing this: a c/h file
pair that reference the struct's field. To keep things low on brittle
code we're generating these two files from libuv_accessor_gen.ml